Update ndSkeletonContainer.cpp
of it is no worth the time tryin to make skeleton solve faster.
There are many place that can be optimized, but it is better to derive the algorithm on paper again rather that
try to modify the existing one.

I will postpone it for next year. after I get the machine learning code going.